refactor(webtransport): non-null + UInt WebTransport error codes, cross-backend exception parity
Make the WebTransport stream-abort exception identical across native (quiche/Apple) and
browser, enforced by a shared commonTest, and fix the error-code modeling per review:

Error-code types (root-cause, end to end):
- QuicStreamAbort.applicationErrorCode: Long? -> non-null Long. Every backend resolves the
  peer's code on a real abort (quiche's out_error_code on ALL THREE bindings incl. JNI; Apple
  via nw_quic_get_stream_application_error), so the nullable was vestigial. It's a 62-bit QUIC
  varint, so stays Long.
- WebTransport layer error code -> UInt (32-bit, draft-ietf-webtrans-http3 §4.3, matching the
  browser's WebIDL unsigned long): socket-http3.WebTransportStreamException.errorCode and the
  new neutral socket-webtransport.WebTransportStreamException.errorCode. WebTransportCloseInfo.code
  and close(code) aligned to UInt too (same 32-bit-unsigned session code).

Cross-backend parity:
- New neutral WebTransportStreamException in socket-webtransport (the browser can't see
  socket-http3's same-named type). Native adapter wraps write-bearing streams to translate
  socket-http3's abort -> neutral (read-side ReadResult.Reset already neutral, passes through).
- Browser backend now MAPS inbound peer aborts: a WebTransportError on read -> ReadResult.Reset,
  on write -> neutral WebTransportStreamException(streamErrorCode) (js + wasmJs). Previously a peer
  reset leaked a raw JS WebTransportError.
- Shared WebTransportTestSuite gains connect_peerResetsStream_surfacesNeutralExceptionWithCode,
  run on JVM (quiche) + Apple (Swift NW) — both GREEN. Chrome harness gains a /reset route +
  browser-side reset assertion (BrowserInteropServer + BrowserWebTransportInteropTest).

Stale-docs/test fixes (the JNI binding was NEVER actually dropping the code):
- quiche_jni.c wires out_error_code into quiche_conn_stream_send and JniQuicheApi reads it back;
  the 'JNI leaves it null' comments were stale (QuicServerTestSuite already asserts the code
  round-trips on every backend). Fixed the QuicheDriver/socket-http3 KDocs accordingly.
- ReactiveDriverTests' assertNull(applicationErrorCode, 'quiche does not surface...') was
  asserting STUB behaviour, not real quiche -> now sets + asserts the code round-trips; StubQuicheApi
  mirrors quiche (out_error_code present, 0-default, only on STREAM_STOPPED/RESET).

Validated: socket-webtransport jvmTest 6/0, AppleWebTransportTest 4/4 (incl. parity), socket-http3
reset round-trip (UInt), ReactiveDriverTests 35/0; all targets compile (jvm/js/wasmJs/apple) + ktlint clean.

+++ socket-webtransport/src/jsMain/kotlin/com/ditchoom/socket/webtransport/WebTransportSupportBrowser.js.kt
@@ -82,11 +82,11 @@ internal class BrowserWebTransportSession(
         scope.launch {
             val info =
                 try {
-                    wt.closed.await().let { WebTransportCloseInfo(it.closeCode, it.reason) }
+                    wt.closed.await().let { WebTransportCloseInfo(it.closeCode.toUInt(), it.reason) }
                 } catch (c: CancellationException) {
                     throw c
                 } catch (t: Throwable) {
-                    WebTransportCloseInfo(0, t.message ?: "session error")
+                    WebTransportCloseInfo(0u, t.message ?: "session error")
                 }
             finish(info)
         }
@@ -124,12 +124,14 @@ internal class BrowserWebTransportSession(
         get() = readableStreamFlow(wt.datagrams.readable) { it.unsafeCast<Uint8Array>().asReadBuffer() }
 
     override suspend fun close(
-        code: Int,
+        code: UInt,
         reason: String,
     ) {
         if (isClosed) return
         val info: dynamic = js("({})")
-        info.closeCode = code
+        // WebIDL coerces closeCode to `unsigned long` via ToUint32, so the signed Int bit pattern of a
+        // 32-bit-unsigned code (negative for the top bit) round-trips to the correct unsigned value.
+        info.closeCode = code.toInt()
         info.reason = reason
         try {
             wt.close(info)
